Use get_default_turbulence as fallback
